backendv3/requirements.txt
google-generativeai
langchain
langchain-google-genai
langchain-community
fastapi
uvicorn
python-dotenv
pandas
sqlite-utils
sqlalchemy